# fix #51135 set default quota period to 5ms based on user experience