--- a/cmd/kubeadm/app/util/config/testdata/conversion/master/v1alpha2.yaml
+++ b/cmd/kubeadm/app/util/config/testdata/conversion/master/v1alpha2.yaml
@@ -91,6 +91,7 @@ kubeletConfiguration:
     containerLogMaxSize: 10Mi
     contentType: application/vnd.kubernetes.protobuf
     cpuCFSQuota: true
+    cpuCFSQuotaPeriod: 0s
     cpuManagerPolicy: none
     cpuManagerReconcilePeriod: 10s
     enableControllerAttachDetach: true