@dims 2018-05-04
@derekwaynecarr The `"cpu-cfs-quota"` command line flag in kubelet right? So we would have a `"cpu-cfs-quota-period"` to go along with it?