@derekwaynecarr 2018-05-04
Yes

On Fri, May 4, 2018 at 5:37 PM Davanum Srinivas <notifications@github.com>
wrote:

> @derekwaynecarr <https://github.com/derekwaynecarr> The "cpu-cfs-quota"
> command line flag in kubelet right? So we would have a
> "cpu-cfs-quota-period" to go along with it?
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/63437#issuecomment-386741730>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AF8dbPLiCmJFAt46M0b_94jS34wGFFxYks5tvMohgaJpZM4Ty3li>
> .
>