@dims 2018-05-04
@szuecs looking forward to an update in this PR with the suggestion above