@derekwaynecarr 2018-05-05
see my comment here for why a flag is preferred.  Not all nodes are tuned for latency.

https://github.com/kubernetes/kubernetes/issues/51135#issuecomment-386767390