@szuecs 2018-05-07
@derekwaynecarr I added the cli flag and config option. I hope that I added it everywhere, where it is needed.

I tested, that:
- tests run and report ok % go test ./pkg/kubelet/cm -v 
- I can locally build kubelet: % make WHAT=cmd/kubelet

Let me know if I have to change anything.