@dims 2018-05-08
@szuecs 

- if run "hack/update-bazel.sh" that should take care of the verify job failure. (and bazel-test job failure too)
- the fix for e2e failure should have already gone in, but you may have to rebase to master

So, please squash the commits (for easier review), rebase to master and we should get all green (cross my fingers)

thanks,
Dims