@szuecs 2018-05-09
@dims The update-bazel.sh did no change anything.
 
Without run-in-gopath.sh I get panics all the time, as far as I understood the developer docs I should run it with run-in-gopath.sh. 
I fetched master and merged into the branch, such that is up to date.
I also pulled the godeps with hack/run-in-gopath.sh hack/godep-restore.sh, but it does not change the update-bazel run:

```
[sszuecs@sandor-lab:~/go/src/k8s.io/kubernetes]% hack/run-in-gopath.sh hack/update-bazel.sh
[sszuecs@sandor-lab:~/go/src/k8s.io/kubernetes]% git st
On branch fix/51135-set-saneer-default-cpu.cfs_period
nothing to commit, working directory clean
```

Any hints?