@derekwaynecarr 2018-05-15
I am in favor of this change, but would like to discuss this topic in tomorrow's sig-node meeting in order to reach consensus on the path forward.

/cc @dchen1107