@derekwaynecarr 2018-05-15
@dims - this should not demand a feature gate.