@derekwaynecarr 2018-05-21
@szuecs - i will review the code in depth later today or tomorrow.