@derekwaynecarr 2018-05-24
Hmm, the plot thickens on this PR.

It appears this PR works with the `cgroupfs` cgroup driver and not the `systemd` cgroup driver.

The issue is that when using systemd, systemd only exposes in its dbus API `CPUQuotaPerSecUSec`.

see in runc:
https://github.com/opencontainers/runc/blob/master/libcontainer/cgroups/systemd/apply_systemd.go#L314

and in systemd the period is fixed via constant:
https://github.com/systemd/systemd/blob/81183d9b9934bc508820d7bedb82695c5f74ed27/src/core/cgroup.c#L517

i will reach out to systemd community to see if i am missing something, but we should think about how to proceed with this pr.  if we ever wanted to support a per pod period tuning option, i would want to have that knob settable via systemd, or know its safe to set it behind systemd before merging this since the systemd cgroup driver is GA.

in the interim, we may want to mark this flag as experimental, and disable the ability to update the value when using systemd cgroup driver.