@hjacobs 2018-05-25
@derekwaynecarr @BenPope thanks! @szuecs could you take a look at https://github.com/szuecs/kubernetes/pull/1?