@szuecs 2018-05-28
@BenPope I merged your PR into this PR and squashed the commit. Thanks!