@szuecs 2018-06-01
/test pull-kubernetes-bazel-build