@szuecs 2018-07-04
/test pull-kubernetes-local-e2e-containerized