@szuecs 2018-07-09
@derekwaynecarr @vishh anything that stops this PR from getting a lgtm?
I feel a bit sick of rebasing and fixing merge conflicts and tests, because of refactorings in other places.