@derekwaynecarr 2018-07-26
Can we put this behind a feature gate flag?  The systemd community has prepared a patch to add support for variable quota , but since this option is not yet supported across all cgroup drivers, I would prefer to put it behind a feature gate flag in the interim.