@szuecs 2018-07-26
@derekwaynecarr can you roughly explain where this is done?
I guess it's a  value in a kubelet flag, right?