@dims 2018-07-26
@szuecs you can follow the same pattern as [`SupportPodPidsLimit`](https://cs.k8s.io/?q=SupportPodPidsLimit&i=nope&files=&repos=kubernetes)