@szuecs 2018-07-31
I added feature gate CustomCPUCFSQuotaPeriod to enable the feature. 
I will check tomorrow, if the tests are passing.