@szuecs 2018-08-05
/test pull-kubernetes-e2e-gke