@szuecs 2018-08-06
/test pull-kubernetes-e2e-gke
/test pull-kubernetes-e2e-gce
/test pull-kubernetes-e2e-gce-100-performance