@dims 2018-08-09
/test pull-kubernetes-e2e-gke
/test pull-kubernetes-cross
/test pull-kubernetes-e2e-kubeadm-gce
/test pull-kubernetes-local-e2e

Run the skipped tests just to be sure!