@szuecs 2018-08-09
@derekwaynecarr PR is ready for review, feature gates set up, more unit tests added, all e2e tests green.