@szuecs 2018-08-14
/test pull-kubernetes-integration