@szuecs 2018-08-23
@derekwaynecarr I changed as commented, but see https://github.com/kubernetes/kubernetes/pull/63437#discussion_r212381282, which could be odd. 
I did not add the e2e test, yet.