@derekwaynecarr 2018-08-31
thank you for the the updated test case, i would prefer an e2e test, but i wont hold this pr on that.

if we ever do graduate the feature to beta, it should definitely have tests.

/approve
/lgtm