@dims 2018-09-01
reapplying lgtm from @derekwaynecarr (lost in rebase)

/lgtm